+++ internal/pkg/backli915t/backlight.go
@@ -1,11 +1,11 @@
-package main
+package backli915t
 
 import (
 	"github.com/vdromanov/backli915t/internal/pkg/regs"
 	log "github.com/vdromanov/backli915t/pkg/multilog"
 )
 
-func getBacklightPercent(blcRegContents *int) int {
+func GetBacklightPercent(blcRegContents *int) int {
 	period, cycle := regs.SplitPayload(*blcRegContents)
 	percent, err := regs.CycleToPercent(cycle, period)
 	if err != nil {
@@ -14,7 +14,7 @@ func getBacklightPercent(blcRegContents *int) int {
 	return percent
 }
 
-func setBacklightPercent(percent int, blcRegContents *int) {
+func SetBacklightPercent(percent int, blcRegContents *int) {
 	period, _ := regs.SplitPayload(*blcRegContents)
 	wantedCycle, err := regs.PercentToCycle(percent, period)
 	if err != nil {
@@ -24,8 +24,8 @@ func setBacklightPercent(percent int, blcRegContents *int) {
 	regs.WriteReg(regs.BLC_PWM_PCH_CTL2_REG, payload)
 }
 
-func changeBacklightPercent(value int, blcRegContents *int) {
-	actualBl := getBacklightPercent(blcRegContents)
+func ChangeBacklightPercent(value int, blcRegContents *int) {
+	actualBl := GetBacklightPercent(blcRegContents)
 	newBl := actualBl + value
-	setBacklightPercent(newBl, blcRegContents)
+	SetBacklightPercent(newBl, blcRegContents)
 }
