+++ internal/pkg/backli915t/pwm.go
@@ -1,11 +1,11 @@
-package main
+package backli915t
 
 import (
 	"github.com/vdromanov/backli915t/internal/pkg/regs"
 	log "github.com/vdromanov/backli915t/pkg/multilog"
 )
 
-func getFrequency(blcRegContents *int) int {
+func GetFrequency(blcRegContents *int) int {
 	period, _ := regs.SplitPayload(*blcRegContents)
 	freq, err := regs.PeriodToFreq(period)
 	if err != nil {
@@ -14,8 +14,8 @@ func getFrequency(blcRegContents *int) int {
 	return freq
 }
 
-func setFrequency(frequency int, blcRegContents *int) {
-	initialBacklightPercent := getBacklightPercent(blcRegContents)
+func SetFrequency(frequency int, blcRegContents *int) {
+	initialBacklightPercent := GetBacklightPercent(blcRegContents)
 	_, cycle := regs.SplitPayload(*blcRegContents)
 	log.Debug.Printf("Got cycle: 0x%08x\n", cycle)
 	period, err := regs.FreqToPeriod(frequency)
@@ -25,11 +25,11 @@ func setFrequency(frequency int, blcRegContents *int) {
 	log.Debug.Printf("Got period: 0x%08x\n", period)
 	regs.WriteReg(regs.BLC_PWM_PCH_CTL2_REG, regs.BuildPayload(period, cycle))
 	newBlcRegContents := regs.ReadReg(regs.BLC_PWM_PCH_CTL2_REG)
-	setBacklightPercent(initialBacklightPercent, &newBlcRegContents)
+	SetBacklightPercent(initialBacklightPercent, &newBlcRegContents)
 }
 
-func changeFrequency(value int, blcRegContents *int) {
-	actualFreq := getFrequency(blcRegContents)
+func ChangeFrequency(value int, blcRegContents *int) {
+	actualFreq := GetFrequency(blcRegContents)
 	newFreq := actualFreq + value
-	setFrequency(newFreq, blcRegContents)
+	SetFrequency(newFreq, blcRegContents)
 }
